Automatically download example data
This is the raw data provided from BNL. The citation is printed out
when the data is automatically used.

This HDF5 file is identical to the raw one provided by BNL except
that we added gzip compression to the HDF5 datasets. That reduced
the file size by more than 10x.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>