Set all channels to zero for voxels out of range
If a voxel is out of range in any channel, set all channels to zero
in that voxel.

This is the most intuitive way to handle the ranges, because
otherwise, other channels would dominate and result in misleading
visualizations.

This actually introduces a new way to interactively explore the
data as, by allowing one to hide voxels that are above/below
certain values in a single or multiple channels.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>